외래키 컬럼 타입 불일치 오류 수정

+++ database/migrations/2025_02_25_055355_category_product_table.php
@@ -18,8 +18,8 @@ class CategoryProductTable extends Migration
 //            $table->id();
 //            $table->timestamps();
 
-            $table->integer('category_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->unsignedBigInteger('category_id')->unsigned();
+            $table->unsignedBigInteger('product_id')->unsigned();
 
             $table->foreign('category_id')->references('id')->on('categories');
             $table->foreign('product_id')->references('id')->on('products');
